requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.0.3
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
huggingface-hub>=0.19.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.20.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
sentence-transformers>=2.2.0,<3.0.0

